requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
gunicorn==20.1.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-dotenv==1.2.1
pydantic==2.12.5
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
email-validator==2.1.0
python-decouple==3.8
